CLOSE #6 - Provide extend lib with namespace pattern . - Update tests .
